requirements.txt
streamlit==1.34.0
boto3==1.34.103
pandas==2.2.2
PyPDF2==3.0.1
faiss-cpu
sentence-transformers
